Add Rust package
This adds a Rust package for this language grammar.  This makes it
easier to use the Rust tree-sitter bindings with a statically known set
of languages, since you can just list them as normal Rust dependencies,
instead of having to embed the grammar and set up a build.rs file
yourself.


Former-commit-id: 6a0c7d25ab8a65fe1d1d6ed5c1a8f2ba30a969a9 [formerly b807bf4455d2ce48d36e58f09b3f9bb1026f0e18]
Former-commit-id: 7afbf0c38dc2eec28951f81261ef7d983ec34814